Add files via upload
Tried to lay out the process of how the Hamiltonian is actually going to work in the context of the code.  I think the hardest thing will be to figure out how to have the Pauli operators permute through the tensor products of identity matrices, so we may end up spending a fair amount of time on how best to do it.